requirements.txt
Markdown
Pillow
PySide6
global_hotkeys
mss
numpy
onnxruntime
opencv-python
pyinstaller
pyperclip
transformers
numba